+++ src/pytest_report_extras/extras.py
@@ -8,17 +8,6 @@ from .attachment import Attachment
 from .attachment import Mime
 
 
-# Counter used for image and page source files naming
-count = 0
-
-
-def counter() -> int:
-    """ Returns a suffix used for image and webpage source file naming """
-    global count
-    count += 1
-    return count
-
-
 class Extras:
     """
     Class to hold pytest-html 'extras' to be added for each test in the HTML report.
@@ -34,9 +23,9 @@ class Extras:
             indent (int): The indent to use to format XML, JSON and YAML documents.
             report_allure (str): The Allure report folder.
         """
-        self.images = []
-        self.sources = []
         self.comments = []
+        self.multimedia = []
+        self.sources = []
         self.attachments = []
         self.links = []  # deprecated
         self.target = None
@@ -69,21 +58,24 @@ class Extras:
             escape_html (bool): Whether to escape HTML characters in the comment.
         """
         image, source = self._get_image_source(target, full_page, page_source)
-        self._add_extra(comment, image, source, None, Mime.image_png, escape_html)
+        self.attach(comment, image, source, None, Mime.image_png, "", escape_html)
 
     def attach(
         self,
         comment: str,
         body: str | bytes | dict | list[str] = None,
+        websource: str = None,
         source: str = None,
         mime: str = None,
-        csv_delimiter=',',
+        csv_delimiter: str = ',',
         escape_html: bool = False
     ):
         """
         Adds a step with an attachment to the report.
-        The image is saved in <report_html>/images folder.
-        The webpage source is saved in <report_html>/sources folder.
+        Images are saved in <report_html>/images folder.
+        Webpage sources are saved in <report_html>/sources folder.
+        Videos are saved in <report_html>/videos folder.
+        Other types of files are saved in <report_html>/downloads folder.
         The 'body' and 'source' parameters are exclusive.
 
         Args:
@@ -92,6 +84,7 @@ class Extras:
                 Can be of type 'dict' for JSON mime type.
                 Can be of type 'list[str]' for uri-list mime type.
                 Can be of type 'bytes' for image mime type.
+            websource (str): The filepath of the webpage source.
             source (str): The filepath of the source of the attachment.
             mime (str): The mime type of the attachment.
             csv_delimiter (str): The delimiter for CSV documents.
@@ -101,10 +94,7 @@ class Extras:
             mime = None
         attachment = self._get_attachment(body, source, mime, csv_delimiter)
         mime = attachment.mime if attachment is not None else None
-        if Mime.is_image(mime):  # Add image attachments as 'screenshot' step
-            self._add_extra(comment, attachment.body, None, None, mime, escape_html)
-        else:
-            self._add_extra(comment, None, None, attachment, mime, escape_html)
+        self._add_extra(comment, websource, attachment, mime, escape_html)
 
     def _get_attachment(
         self,
@@ -114,7 +104,7 @@ class Extras:
         delimiter=',',
     ) -> Attachment:
         """
-        Creates an attachment.
+        Creates an attachment from body or source.
 
         Args:
             body (str | bytes | dict | list[str]): The content/body of the attachment.
@@ -135,20 +125,19 @@ class Extras:
                     if self._html:
                         inner_html = decorators.decorate_uri(self._add_to_downloads(source))
                     return Attachment(source=source, inner_html=inner_html)
+                if Mime.is_multimedia(mime) and mime != Mime.image_svg_xml:
+                    return Attachment(source=source, mime=mime)
                 else:
-                    if Mime.is_image(mime):
-                        f = open(source, "rb")
-                        body = f.read()
-                        f.close()
-                    else:
-                        f = open(source, 'r')
-                        body = f.read()
-                        f.close()
+                    f = open(source, 'r')
+                    body = f.read()
+                    f.close()
             except Exception as error:
-                body = f"Error reading file: {source}\n{error}"
-                utils.log_error(None, f"Error reading file: {source}", error)
+                body = f"Error creating attachment from source {source}\n{error}"
+                utils.log_error(None, f"Error creating attachment from source {source}: ", error)
                 mime = Mime.text_plain
-        if Mime.is_not_image(mime) and isinstance(body, bytes):  # Attachment of body with unknown mime
+
+        # Continue processing attachments with body
+        if Mime.is_not_multimedia(mime) and isinstance(body, bytes):  # Attachment of body with unknown mime
             if self._html:
                 inner_html = decorators.decorate_uri(self._add_to_downloads(body))
             return Attachment(body=body, inner_html=inner_html)
@@ -165,6 +154,8 @@ class Extras:
                 body = f"Error encoding HTML body\n{error}"
                 utils.log_error(None, "Error encoding HTML body", error)
                 mime = Mime.text_plain
+        if mime == Mime.image_svg_xml:
+            return Attachment(body=body, source=source, mime=mime)
         return Attachment.parse_body(body, mime, self._indent, delimiter)
 
     def _get_image_source(
@@ -188,60 +179,116 @@ class Extras:
             return None, None
         return utils.get_screenshot(target, full_page, self._fx_sources or page_source)
 
-    def _save_image_source(self, image: Optional[bytes | str], source: Optional[str], mime: str = "image/*"):
+    def _save_image_video_source(self, data: Optional[bytes | str], source: Optional[str], mime: str):
         """
-        When not using the --self-contained-html option, saves the image and webpage source
+        When not using the --self-contained-html option, saves the image or video and webpage source, if applicable,
            and returns the filepaths relative to the <report_html> folder.
         The image is saved in <report_html>/images folder.
+        The video is saved in <report_html>/videos folder.
         The webpage source is saved in <report_html>/sources folder.
         When using the --self-contained-html option, returns the data URI schema of the image and the source.
 
         Args:
-            image (bytes | str): The image as bytes or base64 string.
+            data (bytes | str): The image/video as bytes or base64 string.
             source (str): The webpage source.
-            mime (str): The mime type of the image.
+            mime (str): The mime type of the image/video.
 
         Returns:
-            The uris of the image and webpage source.
+            The uris of the image/video and webpage source.
         """
-        link_image = None
+        link_multimedia = None
         link_source = None
+        data_str = None
+        data_b64 = None
 
-        if isinstance(image, str):
+        if Mime.is_not_multimedia(mime):
+            utils.log_error(None, "Invalid mime type '{mime}' for multimedia content:")
+            return None, None
+
+        if isinstance(data, str):
             try:
-                image = base64.b64decode(image.encode())
+                data_str = data
+                data_b64 = base64.b64decode(data.encode())
             except Exception as error:
-                utils.log_error(None, "Error decoding image string:", error)
-                image = None
-        # suffix for file names
-        index = (0 if self._fx_single_page or (image is None and source is None)
-                 else counter())
-        # Get the image uri
-        if image is not None:
+                utils.log_error(None, "Error decoding image/video base64 string:", error)
+                return None, None
+        else:
+            try:
+                data_b64 = data
+                data_str = base64.b64encode(data).decode()
+            except Exception as error:
+                utils.log_error(None, "Error encoding image/video bytes:", error)
+                return None, None
+
+        if Mime.is_video(mime):
             if self._fx_single_page is False:
-                link_image = utils.save_image_and_get_link(self._html, index, image)
+                link_multimedia = utils.save_file_and_get_link(self._html, data_b64, None, "videos")
             else:
-                mime = "image/*" if mime is None else mime
-                try:
-                    data_uri = f"data:{mime};base64,{base64.b64encode(image).decode()}"
-                except Exception as error:
-                    utils.log_error(None, "Error encoding image string:", error)
-                    data_uri = None
-                link_image = data_uri
-        # Get the webpage source uri
+                link_multimedia = f"data:{mime};base64,{data_str}"
+            return link_multimedia, None
+
+        if mime == Mime.image_svg_xml:
+            if self._fx_single_page is False:
+                link_multimedia = utils.save_file_and_get_link(self._html, data_str, "svg", "images")
+            else:
+                link_multimedia = f"data:{mime};base64,{data_str}"
+            return link_multimedia, None
+
+        if Mime.is_image(mime):
+            if self._fx_single_page is False:
+                link_multimedia = utils.save_file_and_get_link(self._html, data_b64, None, "images")
+            else:
+                link_multimedia = f"data:{mime};base64,{data_str}"
+
         if source is not None:
             if self._fx_single_page is False:
-                link_source = utils.save_source_and_get_link(self._html, index, source)
+                link_source = utils.save_file_and_get_link(self._html, source, None, "sources")
             else:
                 link_source = f"data:text/plain;base64,{base64.b64encode(source.encode()).decode()}"
 
-        return link_image, link_source
+        return link_multimedia, link_source
+
+    def _copy_image_video(self, filepath: str, mime: str) -> Optional[str]:
+        """
+        Copies the image or video and returns the filepath relative to the <report_html> folder.
+        The image is copied into <report_html>/images folder.
+        The video is copied in <report_html>/videos folder.
+        When using the --self-contained-html option, returns the data URI schema of the image/video.
+
+        Args:
+            filepath (str): The filepath of the image/video to copy.
+            mime (str): The mime type of the image/video.
+
+        Returns:
+            The uris of the image/video and webpage source.
+        """
+        if Mime.is_not_multimedia(mime):
+            utils.log_error(None, f"invalid mime type '{mime}' for multimedia file '{filepath}")
+            return None
+
+        data_str = ""
+        if self._fx_single_page:
+            try:
+                f = open(filepath, "rb")
+                data_b64 = f.read()
+                f.close()
+                data_str = base64.b64encode(data_b64).decode()
+            except Exception as error:
+                utils.log_error(None, f"Error reading image/video file '{filepath}'", error)
+                return None
+            return f"data:{mime};base64,{data_str}"
+
+        if Mime.is_video(mime):
+            return utils.copy_file_and_get_link(self._html, filepath, None, "videos")
+
+        if Mime.is_image(mime):
+            extension = "svg" if mime == Mime.image_svg_xml else None            
+            return utils.copy_file_and_get_link(self._html, filepath, extension, "images")
 
     def _add_extra(
         self,
         comment: str,
-        image: Optional[bytes],
-        source: Optional[str],
+        websource: Optional[str],
         attachment: Optional[Attachment],
         mime: Optional[str],
         escape_html: bool
@@ -251,28 +298,26 @@ class Extras:
 
         Args:
             comment (str): The comment of the test step.
-            image (bytes | str): The image as bytes or base64 string.
-            source (str): The webpage source code.
+            websource (str): The webpage source code.
             attachment (Attachment): The attachment.
             mime (str): The mime type of the attachment.
             escape_html (bool): Whether to escape HTML characters in the comment.
         """
         comment = utils.escape_html(comment) if escape_html else comment
+        link_multimedia = None
+        link_source = None
 
         # Add extras to Allure report if allure-pytest plugin is being used.
         if self._allure and importlib.util.find_spec('allure') is not None:
             import allure
-            if image is not None:
-                allure.attach(image, name=comment, attachment_type=mime)
-                if source is not None:
-                    allure.attach(source, name="page source", attachment_type=allure.attachment_type.TEXT)
-
-            elif attachment is not None:
+            if attachment is not None:
                 try:
                     if attachment.body is not None:
                         allure.attach(attachment.body, name=comment, attachment_type=mime)
                     elif attachment.source is not None:
                         allure.attach.file(attachment.source, name=comment)
+                    if websource is not None:
+                        allure.attach(websource, name="page source", attachment_type=allure.attachment_type.TEXT)
                 except Exception as err:
                     allure.attach(str(err), name="Error adding attachment", attachment_type=allure.attachment_type.TEXT)
             elif comment is not None:
@@ -280,13 +325,17 @@ class Extras:
 
         # Add extras to pytest-html report if pytest-html plugin is being used.
         if self._html:
-            if comment is None and image is None and attachment is None:
+            if comment is None and attachment is None:
                 utils.log_error(None, "Empty test step will be ignored.", None)
                 return
-            link_image, link_source = self._save_image_source(image, source, mime)
-            self.images.append(link_image)
-            self.sources.append(link_source)
+            if attachment is not None and Mime.is_multimedia(attachment.mime):
+                if attachment.source is not None:
+                    link_multimedia, link_source = self._copy_image_video(attachment.source, mime), None
+                else:
+                    link_multimedia, link_source = self._save_image_video_source(attachment.body, websource, mime)
             self.comments.append(comment)
+            self.multimedia.append(link_multimedia)
+            self.sources.append(link_source)
             self.attachments.append(attachment)
 
     def _add_to_downloads(self, target: str | bytes = None) -> str:
@@ -299,7 +348,10 @@ class Extras:
         Returns:
             The uri of the downloadable file.
         """
-        return utils.save_file_and_get_link(self._html, target)
+        if isinstance(target, bytes):
+            return utils.save_file_and_get_link(self._html, target, None, "downloads")
+        else:
+            return utils.copy_file_and_get_link(self._html, target, None, "downloads")
 
     # DEPRECATED CODE
     def link(self, uri: str, name: str = None):